+++ packages/chronus/src/apply-release-plan/index.ts
@@ -1,0 +1,1 @@
+export { applyReleasePlan } from "./apply-release-plan.js";
